+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
-All notable changes to `novaunit` will be documented in this file
+All notable changes to NovaUnit will be documented in this file.
 
 ## 1.0.0 - 2020-06-01
 
-- initial release
+- initial release, with tests for Actions, Lenses and Resources.